Address feedback to CopyButton
