+++ apps/postgres-new/components/copy-button.tsx
@@ -1,0 +1,44 @@
+'use client'
+import { useState } from 'react'
+import { Button } from './ui/button'
+import { Check, Copy } from 'lucide-react'
+import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from './ui/tooltip'
+
+export type CopyButtonProps = {
+  value: string
+  className?: string
+  tooltipContent?: string
+}
+
+export default function CopyButton({
+  value,
+  className,
+  tooltipContent = 'Copy migrations',
+}: CopyButtonProps) {
+  const [copied, setCopied] = useState(false)
+
+  const handleCopy = () => {
+    navigator.clipboard.writeText(value)
+    setCopied(true)
+    setTimeout(() => {
+      setCopied(false)
+    }, 1000)
+  }
+
+  return (
+    <>
+      <TooltipProvider>
+        <Tooltip>
+          <TooltipTrigger asChild>
+            <Button variant="outline" size="icon" className={className} onClick={handleCopy}>
+              {copied ? <Check className="h-4 w-4" /> : <Copy className="h-4 w-4" />}
+            </Button>
+          </TooltipTrigger>
+          <TooltipContent>
+            <p>{tooltipContent}</p>
+          </TooltipContent>
+        </Tooltip>
+      </TooltipProvider>
+    </>
+  )
+}
